Introduce base class for future implementations
